+++ kernel/include/devices/ata.h
@@ -71,8 +71,8 @@ public:
 
     static void resetATA(int bus, int drive);
 
-    static bool prepare(int bus, int drive, int command, int num_blocks, int offset); // prepares the device for a command's data packet
-    static bool read(int bus, int drive, uint8_t* buffer, int num_blocks, int offset);
+    static bool prepare(int bus, int drive, int command, size_t num_blocks, int offset); // prepares the device for a command's data packet
+    static bool read(int bus, int drive, uint8_t* buffer, size_t num_blocks, int offset);
     static bool wait(int bus, int drive, int mask, int waitForState);
 
     static std::vector<ATA_Device> findATA();
